(chore) Reorganise server files
